+++ tests/protobuf-net.Grpc.Test/TestBindings.cs
@@ -9,36 +9,33 @@ namespace protobuf_net.Grpc.Test
 {
     class TestServerBinder : ServerBinder // just tracks what methods are observed
     {
-        public HashSet<string> Methods { get; } = new HashSet<string>();
-        public List<string> Warnings { get; } = new List<string>();
-        public List<string> Errors { get; } = new List<string>();
+        public HashSet<string> Methods { get; } = [];
+        public List<string> Warnings { get; } = [];
+        public List<string> Errors { get; } = [];
         protected override bool TryBind<TService, TRequest, TResponse>(ServiceBindContext bindContext, Method<TRequest, TResponse> method, MethodStub<TService> stub)
         {
             Methods.Add(method.FullName);
             return true;
         }
         protected internal override void OnWarn(string message, object?[]? args = null)
-            => Warnings.Add(string.Format(message, args ?? Array.Empty<object>()));
+            => Warnings.Add(string.Format(message, args ?? []));
         protected internal override void OnError(string message, object?[]? args = null)
-            => Errors.Add(string.Format(message, args ?? Array.Empty<object>()));
+            => Errors.Add(string.Format(message, args ?? []));
     }
 
-    class TestChannel : ChannelBase
+    class TestChannel(string target) : ChannelBase(target)
     {
-        public TestChannel(string target) : base(target) { }
         public override CallInvoker CreateCallInvoker()
             => new TestInvoker(this);
 
-        public HashSet<string> Calls { get; } = new();
+        public HashSet<string> Calls { get; } = [];
 
         private void Call(IMethod method)
             => Calls.Add(Target + ":" + method.FullName);
 
-        class TestInvoker : CallInvoker
+        class TestInvoker(TestChannel channel) : CallInvoker
         {
-            public TestChannel Channel { get; }
-            public TestInvoker(TestChannel channel)
-                => Channel = channel;
+            public TestChannel Channel { get; } = channel;
 
             public override TResponse BlockingUnaryCall<TRequest, TResponse>(Method<TRequest, TResponse> method, string? host, CallOptions options, TRequest request)
             {